+++ src/nitrokey/nk3/_device.py
@@ -5,7 +5,7 @@
 # http://opensource.org/licenses/MIT>, at your option. This file may not be
 # copied, modified, or distributed except according to those terms.
 
-from typing import List
+from typing import TYPE_CHECKING, List
 
 from fido2.hid import CtapHidDevice
 
@@ -70,3 +70,9 @@ class NK3(TrussedDevice):
         return cls._list_pcsc_atr(
             list(bytes.fromhex("3B8F01805D4E6974726F6B657900000000006A")), exclusive
         )
+
+
+if TYPE_CHECKING:
+    from contextlib import AbstractContextManager
+
+    _: type[AbstractContextManager[NK3]] = NK3